backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-qdrant
qdrant-client
unstructured[local-inference]
python-dotenv
httpx
# For embeddings (Ollama)
langchain-ollama
